example2/data/requirements.txt
snakemake
numpy
moments-popgen
pandas